add rough API support to projectCards, will refactor
