+++ frontend/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@reach/router": "^1.2.1",
     "immutable": "^4.0.0-rc.12",
+    "axios": "^0.19.0",
     "humanize-duration": "^3.20.0",
     "node-sass": "^4.12.0",
     "react": "^16.8.6",
@@ -46,6 +47,7 @@
   },
   "devDependencies": {
     "combine-react-intl-messages": "^2.0.2",
+    "eslint-plugin-react-hooks": "^1.6.1",
     "react-test-renderer": "^16.8.6"
   }
 }
