+++ frontend/src/components/projectcard/projectCard.js
@@ -14,8 +14,9 @@ function PriorityBox({ priority }: Object) {
     color = "red";
     borderColor = "b--red";
   }
+  const translated = priority ? <FormattedMessage {...messages["projectPriority"+priority]} /> : ""
   return <div className={`pa1 fr w-33 tc br1 mt3 f7 ttu ba ${borderColor} ${color}`}>
-    <FormattedMessage {...messages["projectPriority"+priority]} />
+    {translated}
     </div>;
 }
 
@@ -27,7 +28,7 @@ function ProjectTeaser({ lastUpdated, totalMappers }: Object) {
 
     } else {
         return (<div className="f7 tl blue-light mv2" ><span className="blue-grey b">
-        {totalMappers}</span> <FormattedMessage {...messages["projectTotalContributors"]} />
+        {totalMappers||0}</span> <FormattedMessage {...messages["projectTotalContributors"]} />
         </div>)
     }
   }
@@ -60,20 +61,32 @@ function getLogoClass(organisationTag: String) {
       {className: "org-msf",
       organisationTag: "Médecins Sans Frontières"
     }];
-  //organisationTag: World Bank
-  return orgs.find((a) => a.organisationTag === organisationTag.organisationTag).className
+  if (organisationTag.organisationTag) {
+    return orgs.find((a) => a.organisationTag === organisationTag.organisationTag).className
+  } else {
+    return null
+  }
   
+}
+
+export function MapperLevelLabel({mapperLevel}: Object) {
+  const translated = mapperLevel ? <FormattedMessage {...messages["projectMapperLevel"+mapperLevel]} /> : "";
+  return (
+  <span className="fl f7 mt1 ttc fw5 blue-grey">
+    {translated}
+    </span>);
 }
 
 export function ProjectCard({
   projectId,
-  title,
+  name,
   shortDescription,
   organisationTag,
   lastUpdated,
   dueDate,
+  priority,
   mapperLevel,
-  projectPriority,
+  campaignTag,
   percentMapped,
   percentValidated,
   totalMappers
@@ -82,21 +95,21 @@ export function ProjectCard({
     <a href={`#project=${projectId}`}>
     <article className="fl w-25-l base-font w-50-m w-100 mb3 ph2 blue-dark mw5">
       <div className="pv3 ba br1  b--grey-light ph3">
-        <PriorityBox priority={projectPriority} />
+        <PriorityBox priority={priority} />
         <div className="w-50 red dib"><ProjectOrgLogo organisationTag={organisationTag} /></div>
         <div className="ma1 w-100">
           <div className="f7 blue-grey">#{projectId}</div>
           <h3 className="pb2 f5 fw6 h3 lh-title overflow-y-visible">
-            {title}
+            {name}
           </h3>
           <div className="tc f6">
             <div className="w-100 tl pr2 f7 blue-light dib lh-title mb2 h2 overflow-y-hidden">
-              {shortDescription}
+              {shortDescription} {campaignTag ? " · " + campaignTag : ""}
             </div>
             <ProjectTeaser totalMappers={totalMappers} lastUpdated={lastUpdated} />
             <ProjectProgressBar percentMapped={percentMapped} percentValidated={percentValidated} />
             <div className="cf pt2 h2">
-              <span className="fl f7 mt1 ttc fw5 blue-grey"><FormattedMessage {...messages["projectMapperLevel"+mapperLevel]} /></span>
+              <MapperLevelLabel mapperLevel={mapperLevel} />
               <DueDateBox dueDate={dueDate} />          
             </div>
           </div>
